+++ R/assertthat.R
@@ -62,8 +62,8 @@ assertthat::on_failure(is_nonnegative_count) <- function(call, env) {
 #' is_variables(list(1, 2))
 #' is_variables(list("bla"))
 is_variables <- function(x) {
-  utils.nest::is_fully_named_list(x) &&
-    all(vapply(x, utils.nest::is_character_vector, TRUE))
+  checkmate::check_list(x, names = "named") &&
+    all(vapply(x, checkmate::check_character, TRUE))
 }
 assertthat::on_failure(is_variables) <- function(call, env) {
   paste0(deparse(call$x), " is not a list of variable names")
@@ -203,7 +203,7 @@ assertthat::on_failure(is_equal_length) <- function(call, env) {
 #' is_proportion(x = 1.3)
 #' is_proportion(x = 0, include_boundaries = TRUE)
 is_proportion <- function(x, include_boundaries = FALSE) {
-  utils.nest::is_numeric_single(x) &&
+  checkmate::check_numeric(x, len = 1) &&
     is_proportion_vector(x, include_boundaries = include_boundaries)
 }
 assertthat::on_failure(is_proportion) <- function(call, env) {