+++ R/h_response_subgroups.R
@@ -111,10 +111,10 @@ h_proportion_subgroups_df <- function(variables,
     is.character(variables$rsp),
     is.character(variables$arm),
     is.character(variables$subgroups) || is.null(variables$subgroups),
-    utils.nest::is_character_single(label_all),
     is_df_with_variables(data, as.list(unlist(variables))),
     is_df_with_nlevels_factor(data, variable = variables$arm, n_levels = 2)
   )
+  checkmate::assert_string(label_all)
 
   # Add All Patients.
   result_all <- h_proportion_df(data[[variables$rsp]], data[[variables$arm]])
@@ -327,10 +327,10 @@ h_odds_ratio_subgroups_df <- function(variables,
     is.character(variables$arm),
     is.character(variables$subgroups) || is.null(variables$subgroups),
     is.character(variables$strat) || is.null(variables$strat),
-    utils.nest::is_character_single(label_all),
     is_df_with_variables(data, as.list(unlist(variables))),
     is_df_with_nlevels_factor(data, variable = variables$arm, n_levels = 2)
   )
+  checkmate::assert_string(label_all)
 
   strata_data <- if (is.null(variables$strat)) {
     NULL