+++ R/survival_time.R
@@ -48,9 +48,9 @@ s_surv_time <- function(df,
   assertthat::assert_that(
     is_df_with_variables(df, list(tte = .var, is_event = is_event)),
     assertthat::is.string(.var),
-    utils.nest::is_numeric_vector(df[[.var]]),
-    utils.nest::is_logical_vector(df[[is_event]])
   )
+  checkmate::assert_numeric(df[[.var]], min.len = 1)
+  checkmate::assert_logical(df[[is_event]], min.len = 1)
 
   conf_type <- control$conf_type
   conf_level <- control$conf_level