+++ tests/testthat/test-table_vst01.R
@@ -49,7 +49,7 @@ gen_advs <- function() {
 }
 
 testthat::test_that("VST01 default variant is produced correctly", {
-  utils.nest::skip_if_too_deep(3)
+  skip_if_too_deep(3)
 
   advs <- gen_advs()
   advs_baseline <- advs %>% dplyr::filter(ABLFL == "Y") # nolint